Supports Fertilizer items also
